Set conmon scope KillSignal to SIGPIPE
conmon forwards the SIGTERM to the main container process.
This is undesirable during node shutdown as it results in
the container receving 2 SIGTERMs. With this change, conmon
ignores the SIGPIPE and will exit after the main container
process dies. If it doesn't for some reason, then
it gets SIGKILL by systemd after 90 seconds.

We need a follow-on to adjust the 90s to a value higher
than the termination grace period as a follow on. This doesn't
matter today as we don't retain container state on reboot.

Signed-off-by: Mrunal Patel <mrunalp@gmail.com>