oci: check process is still around with kill
we previously were verifying the proc entry was enough to ensure the process was still around
unfortunately, if the parent doesn't call waitpid, that process isn't cleaned up.
that may cause the process to seem to be around, when it acutally is no longer running.

Running Kill() isn't foolproof either, as it still races a bit, but it improves the situation by protecting
against a very slow waitpid

Signed-off-by: Peter Hunt <pehunt@redhat.com>