inspect: send container PID for dropped infra sandbox
cAdvisor uses the container PID to get the network stats from /proc/$pid/net/dev.
When we drop the infra container, there isn't a PID associated with the infra container.

Since we use the same network namespace for a pod, we can return the first container's PID as the PID, and we will
get the network stats for the pod.

Signed-off-by: Peter Hunt <pehunt@redhat.com>