fix: Don't set `image-endpoint` in crictl config
`crictl` sets its path for the image management RPC endpoint to the same
value as the runtime endpoint if it is not otherwise specified. This
means that specifying both to be the same path is technically redundant.
While this is fine for an environment with a single container runtime
to be interacted with, it makes things a little more complicated when
the user attempts to communicate with an alterate runtime by only
specifying a `runtime-endpoint` path and relying on the documented
behaviour of `crictl`.

This change removes the `image-endpoint:` line from the `crictl.yaml`
file in this repository which gets bundled into the cri-o RPMs. Doing so
ensures that as long as cri-o uses the same path for the runtime and
image RPC endpoints, a user can do the naive thing and execute
`crictl --runtime-endpoint /somewhere/else [command]`, assuming thatthe
path to the alternate runtime's RPC will be used as the image endpoint
for that command too.

Signed-off-by: Matt Brindley <58414429+maybe-sybr@users.noreply.github.com>