stats: fix log spam
Because we don't lock in stats collection (to keep it speedy), we run into
races with other routines deleting stuff from underneath us.

As such, we should be wary to print errors in these expected situations.

Instead, ignore or reduce the log level for some common errors that are not really worth reporting.

Signed-off-by: Peter Hunt <pehunt@redhat.com>