Improve image name tag and digest retrieval
The image result features a name and digest which can be re-used for the
image list result. This means we do not return `<none>:<none>` and
`<none>@<none>` any more in case we do not have any name available. This
aligns with the Kubernetes scheduler which takes the result of the image
list into account.

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>