refactor handling of allowed_annotations
The code needed to support a new allowed_annotation was both difficult to maintain and inefficient.

instead of checking we allow an annotation (by searching through a vector each time),
filter out the disallowed annotations of the container and sandbox configs.

Now, if the container or pod still has the annotation, we know we can operate on it

Signed-off-by: Peter Hunt <pehunt@redhat.com>