crio wipe: only completely wipe storage after a reboot
We created the CleanShutdownFile to work around issues where nodes didn't have time
to sync changes to disk before rebooting. This caused an unclean reboot.

However, for this to be relevant, the node had to actually reboot. If it didn't,
we could fail to remove the storage because the containers are still running.

In this case, we shouldn't wipe anything.

Signed-off-by: Peter Hunt <pehunt@redhat.com>