server: add support for absent_mount_sources_to_reject
To match docker behavior, a directory is created when the src of a mount does not exist.

This causes issues when the src should not be a directory (/etc/hostname is a prime example).

Instead of relying on correct user behavior, we should urge them to do the right thing by erroring in this case (if a configured mount source is added but does not exist)

Signed-off-by: Peter Hunt <pehunt@redhat.com>