storage: use volatile containers
overlay supports a new mount option "volatile" that reduces I/O by
ignoring fsync and syncfs requests.

Enable it for CRI-O containers as they are "volatile" by nature and
are not supposed to survive a reboot.

Even in the case containers created by CRI-O are persistent, we
already have code in place to throw away the storage if the shutdown
was not clean, so they will be cleaned up in any case.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>