config,runtime: Add runtime_config_path option
`runtime_config_path` is a "VM" runtime type specific entry, which
allows the admin to set an arbitrary configuration file to be used to
spawn a container runtime.

This will help to simplify how kata-deploy currently sets different
configuration files for kata-containers*, and also give us the ability
to easily provide any special configuration to a customer or so.

*: In case you're interested, it's currently done by a shell script
wrapping the call to the containerd-shim-v2, setting the configuration
file via an environment variable.  Definitely not the best case scenario.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>