test/test_runner: rm eval, fix comment
In order to correctly process spaces in argument, we need to use arrays,
and we can't use eval (as it splits arguments by spaces).

In the current code, eval is needed precisely for the reason of
splitting the arguments by spaces.

This splitting, though, can be avoided if we use bash arrays.
This is what the code does.

Another issue is, when we remove "eval", bash builtin time is no longer
used, so a time binary is required, which is not available on some CI
environments. The fix to this is to move time into execute(), so now
bash recognizes it and uses the builtin.

While at it, make it more obvious that the default value for TESTS is
"." (a current working directory).

This fixes the following shellcheck 0.8.0 warning:

	In test/test_runner.sh line 25:
	    eval "$@"
		  ^-- SC2294 (warning): eval negates the benefit of arrays. Drop eval to preserve whitespace/symbols (or eval as string).

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>