pinns: Pass sysctls as repeated '-s' arguments
Use repeated '-s' arguments to pass sysctls keys and values to pinns
to replace the serialization in cri-o and parsing in pinns.

To simplify the logic, this allocates an 'argc/2' sized array to store
the pointers to the sysctls `k=v` pairs. This array will in most cases
be much bigger than what is really needed but this should be safe as
argc is limited in size on Linux.

See: https://github.com/cri-o/cri-o/security/advisories/GHSA-6x2m-w449-qwx7

Signed-off-by: Timothée Ravier <tim@siosm.fr>