Add scorecards Github action
Adds the Github scorecards action
https://github.com/ossf/scorecard-action

The only limitation is that there is no Personal Access Token (PAT)
attached as I cannot generate this for the repo.

The scorecard action is focused on highlighting security of a repository
and its dependencies. It includes various heuristics and scanning checks
which can be helpful in assessing and guiding the security posture of
cri-o.

Signed-off-by: David Korczynski <david@adalogics.com>