Bump golang dependency on containerd
Updating containerd to 1.6.4.
Had to modify the way we call the Command() function from containerd:
- use of the CommandConfig structure as parameter
- use the actual binary name
  (no more converting to containerd binary naming pattern)

This later change was needed as in Containerd 1.6, the validation of the given
binary path is not done in Command() anymore - the path is used as is, allowing
custom runtime paths.
(see https://github.com/containerd/containerd/commit/6870f3b1)

Signed-off-by: Julien Ropé <jrope@redhat.com>