oci: save criStatus similarly to criContainer
when a message is being sent over proto, it must not be edited. To circumvent this, the underlying memory section is not updated,
but the whole memory address with a changed copy.

Do this similarly for criStatus, which is a bit more involved, as there are more changing pieces.

In order to do so, we introduce Set{Started,Stopped} calls and wire them into runtime_oci.go. This way, we can return an updated
criStatus, but atomically change it so as to not corrupt proto's view of the world

Signed-off-by: Peter Hunt <pehunt@redhat.com>