Inherit capabilities for container exec's
We dropped the inheritable capabilities support for the container
creation in 3e9d7725714057d5fc2ce7eb9f905dc50c96647c to fix a security
vulnerability. This had the side effect that all exec's within
containers break the existing behavior if they rely on the initially set
capabilities, because the exec process does not inherit the capabilities
of the main container any more. We now explicitly set the inheritable
capabilities for all exec's to restore the changed behavior.

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>