highperfhooks: avoid unbound growth of irqbalance
The highperformance hooks can update the irqbalance
configuration file to adjust the current banned cpu list,
depending on pod annotations. This is done to make sure
that guaranteed pods that declare they don't want to have IRQ
processing on the CPUs they're assigned to, gets that.

When the update is done in-place, so likely most of the time,
a stray '\n' is added on _every_ update. The single character
always adds up, creating a unbounded growth of the configuration
file.

In practical terms, this is likely not really too bad, because
we will need to get in the ballpark of thousands updates before
the file size starts becoming concerning.

Nevertheless, add the (trivial) fix and a test to cover this scenario.

Signed-off-by: Francesco Romani <fromani@redhat.com>