oci: take opLock for UpdateContainer
UpdateContainer requests can come in at any time in a container's
lifecycle, even when it's shutting down.

In some cases, it can cause a container to get stuck: runc pauses the
cgroup on v1 and systemd to prevent systemd from mucking around while
the cgroup is being updated. In some cases, CRI-O calls
`UpdateContainerStatus` during that time, and gets `paused` state. CRI-O
does not expect a container to be paused, so it doesn't know how to
handle it.

Fix this by taking opLock when updating

Signed-off-by: Peter Hunt <pehunt@redhat.com>