irqbalance: fix padding of odd-sized cpu mask
The high performance hook support the dynamic IRQ balancing thanks some
well known annotation.
Pods can opt out from IRQ handling exposing this annotation. On pod startup,
crio will recompute the irqbalance ban mask and reapply irqbalance to make
sure the cpus assigned to these pods will not handle IRQs - being added to
the irqbalance ban mask.
This works only for guaranteed pods requesting integral CPUs.

To support this feature, crio needs to recompute the cpu ban mask.
In turn, crio depends on encoding/hex.DecodeString for some of the heavy lifting.
The function expects to deal with even-sized inbound strings (which are trivial
transformations of the irqbalance ban mask).

So crio does left-aligned zero padding of the irq affinity mask at the beginning
of computation.

The left-padded mask is then adjusted excluding any relevant cpu, and then inverted.
Here lies the problem: the padding character should always be "0" (ASCII zero), not
"f", which is obtained inverting the original correct zero character.

The outcome of this behavior is that stray 'f' end up in the IRQ ban list in the
irqbalance config file, and they are never removed.

Fixing the padding or the mask inversion require relatively invasive
code changes, for example to pass around the expected mask length and
act accordingly. While not particularly challenging per se, these change
don't fit smoothly in the current code.

Another option is to clamp the computed masks to the expected length
once the computation and the inversion has been done. We pursue this
approach in this change.

Fixes: https://github.com/cri-o/cri-o/issues/6145
Signed-off-by: Francesco Romani <fromani@redhat.com>