Track type of all bind mounts during checkpointing
This change is especially helpful to restore containers checkpointed in
Kubernetes.

During restore CRIU remounts all bind mounts which have been mounted in
the container during checkpointing. If the source of the bind mount is
missing, because the restore is on another system or if the restore
happens outside of Kubernetes and Kubernetes specific mount points are
missing, CRIU will fail during restore.

Interestingly CRI-O creates all missing bind mount sources during
container start automatically. During container restore this is
disabled, as CRI-O creates all missing bind mounts as directories.

One very prominent case where this produces an error during restore, is
/etc/resolv.conf. CRI-O would create the source of this bind mount as a
directory and CRIU will try to mount the just created directory as on
the container internal file /etc/resolv.conf. This will fail as the
kernel does not bind mount a directory on a file.

If the container is restored as part of Kubernetes, then Kubernetes will
create all the usual bind mount sources and tell CRI-O to create/restore
the container with those new mount points. CRI-O will adapt the
checkpointed container to mount the bind mounts from the new location.

If the container is restored outside of Kubernetes this information is
missing. CRI-O will tell CRIU to restore the container, CRIU will try to
remount the bind mount and fail if it is missing. The user will get an
error message and needs to create the missing external bind mount source
and then try the restore again. To avoid situations like this, CRI-O
will now track all external bind mount sources and the type and recreate
the external bind mount sources to make the restore successful.

Also reduces the time required for the checkpoint/restore tests from 30
seconds per test case to 10 seconds.

Signed-off-by: Adrian Reber <areber@redhat.com>