nri: add experimental NRI adaptation interface.
Implement an NRI adaptation interface. This takes care
of relaying requests/respones to/from NRI (external NRI
plugins) as well as the logic of applying NRI-requested
container adjustments/updates to actual containers.

This interface is unaware of the internal details of
how we represent pods and containers. The functions to
deal with that, the 'NRI Domain' should be implemented
by server.Server.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>