config: do not remove runc if different default runtime
This commit reverts d2ded1d73c9ce282028ec2b4d6016546267322a9 and adds a
test that maintains what it was trying to do.

When upgrading from a cluster that uses runc to one that uses crun,
there may be workloads that specifically request the old default. Those
workloads will fail on restart. To fix this, don't remove runc from the
map. We've since added code that means we don't fail if one of the
non-default runtimes fails validation (if the binary doesn't exist) so
we can handle that situation, which
d2ded1d73c9ce282028ec2b4d6016546267322a9 meant to fix.

Signed-off-by: Peter Hunt <pehunt@redhat.com>