test: config tests should ignore default runtime environment variable
Config tests are meant to verify the configuration code, by exercising
it in specific configurations.
When run in an environment where the default runtime is modified, these
tests fail because the configuration they use do not have this runtime
defined.
As the runtime itself is not involved in these tests, it seems safe to
unset the CONTAINER_DEFAULT_RUNTIME variable in their context.

Signed-off-by: Julien Ropé <jrope@redhat.com>