Fix quoting issue in usage help text
When using the backtick/grave (`) in the usage string for flags as a
quote symbol, the quoted text is used as the placeholder and the
instance in the usage string becomes unquoted.

Use single quote (') instead where this placeholder processing
behavior is not intended.

Implementation of this behavior is from the urfave/cli package, see:
https://github.com/cri-o/cri-o/blob/a2da00682712bf72a7a4ca5fd53da443c20a297e/vendor/github.com/urfave/cli/v2/flag.go#L228-L243

Signed-off-by: Caleb Xu <calebcenter@live.com>