Removed pod checkpointing support
The pod checkpointing code was a left over from the initial merge and
the initial proof of concept about checkpointing and restoring pods.

With the way the CRI ContainerCheckpoint() is defined this code cannot
be used and needs to be removed.

For actual pod checkpointing it needs to be implemented differently
depending on how the corresponding CRI calls might be designed.

Signed-off-by: Adrian Reber <areber@redhat.com>