Add additional metadata to inspect and checkpoint
This commit adds additional metadata to the inspect command if
checkpointing support is active. The following three fields are
added to inspect:

  "checkpointedAt": timestamp,
  "restored": true or false

With "checkpointedAt" it is possible to see when a checkpoint was
created. "restored" is true if the container has been restored from a
checkpoint for all other container this is set to false.

Signed-off-by: Adrian Reber <areber@redhat.com>