internal/factory/containers: allow non-linux containers
This uses the build's GOOS value to generate.New, allowing it to create
a FreeBSD generator in FreeBSD builds. In future, we should allow this
to be overridden since FreeBSD supports Linux emulation - in podman, we
use the image OS value.

We also avoid dereferencing config.Linux if its nil which is the case on
FreeBSD.

Signed-off-by: Doug Rabson <dfr@rabson.org>