Use ImageRef instead of ImageName for restore
Up until now CRI-O was using ImageName to remember the base image of the
restored container. This works, but has two problems:

 * The output from the restored container was different, especially for
   the Kubernetes use case.
   The output of 'crictl ps' contains for the original container under
   'IMAGE' something like 'registry/path/container@sha256:123444444...'.
   The restored container was, however, only displaying something
   like 'registry/path/container'.

 * More problematic, however, was that CRI-O might pull the wrong image
   from the registry.  If the container in the registry was updated (new
   latest tag for example) all of a sudden the wrong base image would be
   downloaded and the restore would potentially fail.

With this change the output of Kubernetes containers in 'crictl ps'
'IMAGE' does not change after restoring the container and the restored
container will run on exactly the same base image.

Signed-off-by: Adrian Reber <areber@redhat.com>