runtimeVM: notify server that the container exited
runtimeVM is monitoring the execution of the container on its own,
without using conmon. Because of that, when the container exits,
the processing is different than for regular containers.
This is causing some issues, like events not being generated on time.

This commit makes runtimeVM create a file under the "ContainerExitPath"
in the same way that conmon does, so that the Server.monitorExits() function
can pick it up and run the required processing for those containers too.

Signed-off-by: Julien Ropé <jrope@redhat.com>