Setting up a per-container imageserver mapping.
A new ImageServerList type is created, which is a map of ImageServers to containerIDs.
Now rather than one single instance of ImageServer, we have possibly one per container.

ImageServer can be registered along with their runtime implementation.
The goal is to provide a way for some runtimes (e.g: peer-pods use case) to process
the Image requests by relaying them to a remote VM, rather than having them
executed on the local host.

Signed-off-by: Julien Ropé <jrope@redhat.com>