test: add a runtime command line tool for kata containers
The "runtime()" helpper command is used to make direct calls to the runtime binary.
This command fails for kata containers, where there is no such tool to call.

This commit adds a helps script that mimics the behaviour of the different
calls to the "runtime()" function.
The runtime() function is modified to call this script when kata containers
are used.

Signed-off-by: Julien Ropé <jrope@redhat.com>