Drop port-opening code from hostport manager
This was copied from the original upstream kubernetes hostport code,
but upstream has since given up on this idea as unworkable (and, in
particular, no longer does port opening for NodePorts in kube-proxy,
so doing it for HostPorts here does not do anything to prevent
HostPort/NodePort conflicts any more).

The scheduler already guarantees that there are no hostport overlaps
between pods on a node, and the cluster should just be configured to
not allow overlaps between HostPorts, NodePorts, and non-Kubernetes
services.

Signed-off-by: Dan Winship <danwinship@redhat.com>