test: use 'runtime state' to retrieve the bundle of a container
In cgroups.bats, when looking up the bundle value of a container, the
test was using "runtime list", parsing the output to retrieve the
expected column.

This relies on runc not changing the output format. It also makes it
confusing as to what value we're actually looking for here.

As other tests are relying on the "runtime state" call, using json output
and 'jq' to parse it, it seems more reliable to do the same here and avoid
possible confusion.

Signed-off-by: Julien Ropé <jrope@redhat.com>