tests: add a "runtime_pid" wrapper
Getting the PID associated with a container by using the
"runtime()" wrapper is different for runc and kata runtimes.
As the kata version is just a custom script, we don't have a
json output that we can parse with jq.

This commit adds another wrapper called "runtime_pid" which relies
on the runtime() wrapper, but processes the output differently for
regular and "vm" type runtimes.
The output will be the PID for the give container, making it
possible to run the same tests for both runtime types transparently.

Signed-off-by: Julien Ropé <jrope@redhat.com>