Don't use KUBE-MARK-MASQ in hostport rules
This is considered "private API" of kube-proxy and in particular, as
of kube 1.27, it is now only created by kube-proxy, not by kubelet, so
it can't be used on systems with non-kube-proxy service proxies.

(Even worse, cri-o was creating the KUBE-MARK-MASQ chain itself if it
didn't already exist, but assuming kubelet would fill in the rules, so
it doesn't even fail in an obvious way in this case.)

The new code avoids the need for using a reserved packet mark bit by
just re-matching the packets that need masquerading at POSTROUTING
time rather than marking them when it DNATs them and then matching the
mark at POSTROUTING time.

Signed-off-by: Dan Winship <danwinship@redhat.com>