tests: add a fake pinns call to delay things and cause the timeout
The test for pod creation timeout did not trigger, because the creation
time was too quick.
Using the same mechanism as "conmon", creating a fake "tmp_pinns" script
that delay the call to the actual pinns binary.

This ensure that the pod creation is too long, and the test succeeds.

Signed-off-by: Julien Ropé <jrope@redhat.com>