sandbox: Handle PodLinuxOverhead and PodLinuxResources CRI fields
The CRI PodLinuxOverhead and PodLinuxResources fields were not saved
which meant that NRI plugins would need to cache the data if
cri-o is restarted. These two fields are already passed to NRI plugins
by containerd. Solve this issue by caching those two fields
and restoring them if cri-o is restarted.

Signed-off-by: Jukka Rissanen <jukka.rissanen@linux.intel.com>