server: do not take lock to populate pid in container status and inspect
We can do this best-effort, and don't need to take the lock to do so.
Also, by using the InitPid, we're not prone to reading a variable that changes underneath us,
as the init pid is set once and not changed.

for inspect, we can use Created without a lock because it's only set once.

Signed-off-by: Peter Hunt <pehunt@redhat.com>