test/README.md: fix wrong bats example
In the example, the variable $status is never set (as it is only set by
"run" which is not used here), so the checking code is not correct.

The correct way to check that the command exited with status of 0 is not
to do anything (as bats runs tests with "set -e", which will fail the
test in case any command will return non-zero exit code).

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>