test/cdi.bats: rm run_cmd, use run
Since bats 1.5.0, run can accept optional exit code (see [1]).

Basically, "run_cmd" reimplements "run", so we can just replace one with
another and be done with it.

This is what the commit does, but in addition, we

1. Simplify the code like

	run_cmd -0 crictl create ...
	ctr_id="$output"

to

	ctr_id=$(crictl create ...)

which has the same effect (thanks to "set -e" implicitly set by bats, it
does check the exit code of the command inside the $(), and that's all
we need here).

2. Simplify the code like

	run_cmd -0 cmd

to

	cmd

(again, thanks to "set -e", we have implicit exit code checking).

[1]: https://bats-core.readthedocs.io/en/stable/writing-tests.html#run-test-other-commands

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>