test/nri.bats: fix checking exit code
These tests run $NRITEST_BINARY and when check $status env var.

The problem is, $status is only set by bats' "run" command.

The proper way to check that the binary exited with 0 is do to nothing
(as bash checks exit code to be 0 thanks to "set -e" implicitly set by
bats).

Fixes: 54b7b5fc0943f94f45aa0d7fb7fe28051ad93351
Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>