test: simplify non-zero exit status checks
Instead of using

	run cmd
	[ "$status" -ne 0 ]

since bats 1.5.0 we can use a simple form:

	run ! cmd

Alas, bats has a bug which makes it change the value of "i" when run is
used with any options ([1], [2]). To workaround that, we should not use
"i" in any code which uses "run" with options. I know, this is crazy.

[1] https://github.com/bats-core/bats-core/issues/726
[2] https://github.com/bats-core/bats-core/pull/727

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>