runtime handler hooks: add DefaultCPULoadBalanceHooks
disabling cpu load balancing requires all overlapping cpusets to have the
cpuset.sched_load_balance field to 0. There is a bug where stale containers
that have stopped still have the field set to 1, and cpumanager won't attempt to
load balance the cpus away from them. Thus, adding a pod that needs cpu load
balancing to be disabled after other pods have stopped causes the cpu load balancing
to not apply in the kernel.

Fix this by adding PostStop hooks for all containers (when high performance hooks are enabled).
This hook will disable sched_load_balance in the container's cgroup. This cgroup is stale, as
the containers is already stopped and won't be restarted, so this won't change behavior past
allowing cpu load balancing to be disabled for new containers.

Signed-off-by: Peter Hunt <pehunt@redhat.com>