test/metrics oom: try to fix
Sometimes OOM is not happening in our tests.

It can be due to swap (so disable it, by setting mem+swap=mem).

It can also be because dd does not increase memory usage over time,
is using too much CPU, and uses kernel devices and syscalls.

So, replace it with a pure user-space memory eater in bash.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>