test/ctr: fix "should not leak children" test case
The test was getting ALL conmon PIDS, which doesn't really work in case
we run tests in parallel (which we do), or in case conmon-rs is used.

In fact, what we need is PIDs of all processes in the container. The
best place to get these is cgroups.

Do that, and simplify process status check using grep.

Fixes: 9fea6e219e
Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>