Vendor in stable containers/storage instead of stray commit
We may or may not be seeing an issue related to this in kubevirt on OpenShift,
but regardless, it doesn't seem like a bad idea to vendor in a stable release instead
of a stray commit. We also vendor this exact version in 1.26 at the point of writing this.
Diff:
https://github.com/containers/storage/compare/a747b27...v1.45.3

The issue we are seeing is
```bash
sh-4.4# cat /proc/1/mountinfo   | grep 'ddd0718f846fa69a311121696360a9d6f2e2d9258ffc40'
sh-4.4# cd /var/lib/containers/storage/overlay/ddd0718f846fa69a311121696360a9d6f2e2d9258ffc407cee3fd9b992122ebb
sh-4.4#
```
meaning, the overlay path exists, but there is no such mount
the mount was succeeded at one point according to the journal,
so it's weird to me that it's not in /proc/1/mountinfo.

Signed-off-by: Alex Kalenyuk <akalenyu@redhat.com>