Update `README.md` version table
We always forget to add the new release, which is now not required any
more considering a new format. We also outline that the patch release of
CRI-O (`.y`) does not have to match with them from Kubernetes (`.z`).

Since we only have two data rows within the table, we also can avoid the
key mapping and add their description directly to it.

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>