test/metrics: simplify oom test, add debug
Currently, this test case keeps waiting for the container that was
already killed in case OOM detection has failed. This is a huge (15
minutes) waste of time.

To fix, change the logic to wait for container exit, and then check the
OOM. This also makes the test code more readable.

Also, add some debug info in case the test has failed.

Also, add a sleep before the memory eater to give the container some
time to live.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>