Add Kubernetes package test for static binary bundle
This test ensures a basic working setup for kubeadm bootstrapped
clusters using the official packages.

We also have to add a default `registries.conf` to the bundle to make it
work. As follow-up we will remove non-required binaries like the CNI
plugins (no installed in parallel to the official packages).

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>