Remove call stateCmd twice.
There are two situations that cause stateCmd to return the container state as ContainerStateStopped. One is when calling the OCI state interface, which returns the current container state as ContainerStateStopped, and the other is when calling the OCI state interface encounters an error, forcing the container state to be set as ContainerStateStopped.

1. In the first situation, if stateCmd returns the container state as ContainerStateStopped, running it for the second time will still result in ContainerStateStopped, because the container cannot recover from ContainerStateStopped to running or any other state. The container has only four states: Created, Paused, Running, and Stopped. Stopped is already the final state.
2. In the second situation, if there is an cmd error, canReturn will skip the subsequent second call. The subsequent second stateCmd code will not be executed. Therefore, the second call to stateCmd can be removed.

Signed-off-by: yangyang2 <yangyang2@kylinos.cn>